delete blog route protected done

+++ app/api/blogs/[id]/route.ts
@@ -1,3 +1,4 @@
+import { getUserFromToken } from "@/lib/auth";
 import { prisma } from "@/lib/prisma";
 
 export async function GET(
@@ -10,6 +11,15 @@ export async function GET(
     where: {
       id: Number(id),
     },
+    include: {
+      user: {
+        select: {
+          id: true,
+          username: true,
+          email: true,
+        },
+      },
+    },
   });
 
   return Response.json(blog);
@@ -19,7 +29,29 @@ export async function DELETE(
   req: Request,
   { params }: { params: Promise<{ id: string }> },
 ) {
+  const user = await getUserFromToken(req);
+
+  if (!user) {
+    return Response.json({ message: "Unauthorized" }, { status: 404 });
+  }
   const { id } = await params;
+
+  const blog = await prisma.blog.findUnique({
+    where: {
+      id: Number(id),
+    },
+  });
+
+  if (!blog) {
+    return Response.json({ message: "blog not found" }, { status: 404 });
+  }
+
+  if (blog.userId !== user.id) {
+    return Response.json(
+      { message: "Forbidden: You can delete only your own blog" },
+      { status: 403 },
+    );
+  }
 
   await prisma.blog.delete({
     where: {
